--- app/config/services/app.yml
@@ -1,11 +1,0 @@
-# Learn more about services, parameters and containers at
-# http://symfony.com/doc/current/service_container.html
-parameters:
-#    parameter_name: value
-
-services:
-    _defaults:
-        public: false # Making services private improve performances
-
-#    'My\ClassName':
-#        arguments: ["@another_service_name", "plain_value", "%parameter_name%"]
