+++ bin/console
@@ -1,6 +1,7 @@
 #!/usr/bin/env php
 <?php
 
+use App\AppKernel;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Component\Console\Input\ArgvInput;
 use Symfony\Component\Debug\Debug;
@@ -13,7 +14,7 @@ use Symfony\Component\Debug\Debug;
 set_time_limit(0);
 
 /** @var Composer\Autoload\ClassLoader $loader */
-$loader = require __DIR__.'/../app/autoload.php';
+$loader = require __DIR__.'/../conf/autoload.php';
 
 $input = new ArgvInput();
 $env = $input->getParameterOption(['--env', '-e'], getenv('SYMFONY_ENV') ?: 'dev');
