+++ conf/dev/config.yml
@@ -1,9 +1,9 @@
 imports:
-    - { resource: config.yml }
+    - { resource: '../config.yml' }
 
 framework:
     router:
-        resource: "%kernel.root_dir%/config/routing_dev.yml"
+        resource: "%kernel.root_dir%/../conf/dev/routing.yml"
         strict_requirements: true
     profiler: { only_exceptions: false }
