+++ conf/dev/routing.yml
@@ -11,4 +11,4 @@ _errors:
     prefix:   /_error
 
 _main:
-    resource: routing.yml
+    resource: '../routing.yml'
