+++ conf/prod/config.yml
@@ -1,5 +1,5 @@
 imports:
-    - { resource: config.yml }
+    - { resource: '../config.yml' }
 
 #doctrine:
 #    orm:
