+++ conf/routing.yml
@@ -1,0 +1,3 @@
+action:
+    resource: "../src/Controller/"
+    type:     annotation
