+++ conf/test/config.yml
@@ -1,5 +1,5 @@
 imports:
-    - { resource: config_dev.yml }
+    - { resource: '../dev/config.yml' }
 
 framework:
     test: ~
