+++ src/AppCache.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace App;
+
 use Symfony\Bundle\FrameworkBundle\HttpCache\HttpCache;
 
 class AppCache extends HttpCache
