+++ web/app.php
@@ -1,17 +1,12 @@
 <?php
 
+use App\AppCache;
+use App\AppKernel;
 use Symfony\Component\HttpFoundation\Request;
 
 /** @var \Composer\Autoload\ClassLoader $loader */
-$loader = require __DIR__.'/../app/autoload.php';
-if (PHP_VERSION_ID < 70000) {
-    include_once __DIR__.'/../var/bootstrap.php.cache';
-}
-
+$loader = require __DIR__.'/../conf/autoload.php';
 $kernel = new AppKernel('prod', false);
-if (PHP_VERSION_ID < 70000) {
-    $kernel->loadClassCache();
-}
 //$kernel = new AppCache($kernel);
 
 // When using the HttpCache, you need to call the method in your front controller instead of relying on the configuration parameter
